requirements.txt
cffi==1.15.1
cryptography==39.0.1
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
pycparser==2.21
pycryptodome==3.17